# src/app/authentication/docs/swagger.yaml
definitions:
  models.BFFChangePasswordRequest:
    properties:
      confirm_password:
        minLength: 8
        type: string
      new_password:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - confirm_password
    - new_password
    type: object
  models.BFFChangePasswordResponse:
    properties:
      message:
        type: string
    type: object
  models.BFFCreateUserRequest:
    properties:
      confirmPassword:
        example: Secure@123
        minLength: 8
        type: string
      email:
        example: arijit@gmail.com
        type: string
      panCard:
        example: EQZRP1234P
        type: string
      password:
        example: Secure@123
        maxLength: 20
        minLength: 8
        type: string
      phoneNumber:
        example: 7568912340
        maximum: 9999999999
        minimum: 1000000000
        type: integer
      username:
        example: Arijit
        maxLength: 32
        minLength: 5
        type: string
    required:
    - confirmPassword
    - email
    - panCard
    - password
    - phoneNumber
    - username
    type: object
  models.BFFSignInRequest:
    properties:
      password:
        example: Secure@123
        maxLength: 20
        minLength: 8
        type: string
      username:
        example: Arijit
        maxLength: 32
        minLength: 5
        type: string
    required:
    - password
    - username
    type: object
  models.BFFSignInResponse:
    properties:
      message:
        type: string
    type: object
  models.BFFValidateUserOtpRequest:
    properties:
      otp:
        type: string
      username:
        example: Arijit
        maxLength: 32
        minLength: 5
        type: string
    required:
    - otp
    - username
    type: object
  models.BFFValidateUserOtpResponse:
    properties:
      accessToken:
        type: string
      message:
        type: string
    type: object
  models.ErrorAPIResponse:
    properties:
      error:
        type: string
      errors:
        $ref: '#/definitions/models.ErrorMessage'
    type: object
  models.ErrorMessage:
    properties:
      errorMessage:
        type: string
      key:
        type: string
    type: object
info:
  contact: {}
  description: Authentication APIs for Stock Broker Application
  title: Authentication Service API
  version: "1.0"
paths:
  /api/auth/change-password:
    post:
      consumes:
      - application/json
      description: verifies the JWT and user request and changes user's password
      parameters:
      - description: User Change-Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BFFChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password Changed successfully
          schema:
            $ref: '#/definitions/models.BFFChangePasswordResponse'
        "400":
          description: Invalid input payload
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "401":
          description: 'Case A: JWT expired / Case B: JWT invalid / Case C: Unauthorized
            request'
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "404":
          description: User does not exist
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
      security:
      - BearerAuth: []
      summary: Changes user's password
      tags:
      - User
  /api/auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT token
      parameters:
      - description: User Sign In Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BFFSignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Signin successful
          schema:
            $ref: '#/definitions/models.BFFSignInResponse'
        "400":
          description: Invalid input payload
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "404":
          description: User does not exist
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
      summary: Sign in an existing user
      tags:
      - User
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Handles user registration by validating input and storing user
        details
      parameters:
      - description: User Registration Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BFFCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            type: string
        "400":
          description: Invalid input payload
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
      summary: Create a new user
      tags:
      - User
  /api/auth/validate-otp:
    post:
      consumes:
      - application/json
      description: Validates user OTP and return clear success/ failure message
      parameters:
      - description: User OTP Validation Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BFFValidateUserOtpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP validation successful
          schema:
            $ref: '#/definitions/models.BFFValidateUserOtpResponse'
        "400":
          description: Invalid input payload
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "401":
          description: 'Case A: Incorrect OTP / Case B: Expired OTP'
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "404":
          description: User does not exist
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorAPIResponse'
      summary: Validates user OTP
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
    x-extension-openapi: '{"example": "value on a json format"}'
swagger: "2.0"